# api/v2/openapi.yml
openapi: "3.0.2"
info:
  title: Authcore API
  version: "2.0"
paths:
...
  /api/v2/users:
    get:
      summary: List users
      tags:
        - users
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_size:
                    type: integer
                  next_page_token:
                    type: string
                  prev_page_token:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
    post:
      summary: Create a new user.
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verifier:
                  $ref: '#/components/schemas/PasswordVerifier'
                email:
                  type: string
                phone:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  refresh_token:
                    type: string
  /api/v2/users/{id}:
    get:
      summary: Get user
      tags:
        - users
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update user
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                preferred_username:
                  type: string
                email:
                  type: string
                email_verified:
                  type: boolean
                phone_number:
                  type: string
                phone_number_verified:
                  type: boolean
                user_metadata:
                  type: object
                app_metadata:
                  type: object
                is_locked:
                  type: boolean
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      summary: Delete user
      tags:
        - users
      responses:
        "204":
          description: Success
...
  /api/v2/users/{id}/password:
    post:
      summary: Update user password
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordVerifier"
      responses:
        "204":
          description: Success
  /api/v2/users/{id}/roles:
    get:
      summary: Get user's roles.
      tags:
        - user
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    post:
      summary: Assign new role to user
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
...
components:
  schemas:
...
    PasswordVerifier:
      oneOf:
        - $ref: '#/components/schemas/SPAKE2PlusVerifier'
    SPAKE2PlusVerifier:
      type: object
      properties:
        method:
          type: string
        salt:
          type: string
        w0:
          type: string
        l:
          type: string
      required:
        - method
        - salt
        - w0
        - l
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        preferred_username:
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        phone:
          type: string
        phone_verified:
          type: boolean
        user_metadata:
          type: object
        app_metadata:
          type: object
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time